requirements.txt
psycopg2-binary>=2.9,<3.0
pandas>=2.0,<3.0
openpyxl>=3.1,<4.0
click>=8.1,<9.0
